Add support for 'proxynicks'
Proxynicks are those who speak on someone else's behalf.
Typically, they are bots which describe actions taken
by a sysadmin. For WMF wikis, we have logmsgbot, which
gives some !log messages when sysadmins take certain
scripted actions. The information doesn't really come
from the bot, it comes from the user - which is the
first thing in the message. So, we simply take that and
tell the main channel that that's who said the message.